Binding, deep-clone, and index.js added: cloning works!
This is sweet. I've added the binding file, deep-clone.cc, and index.js.
There is a minimal example in index.js showing that the cloning works as
expect, but I'll be adding tests soon.

The core part of deep-clone.cc is the following line:

  `Local<Object> cloned_js_obj = js_obj->Clone();`

From v8's source (v8.h), Clone() creates a "fast but shallow copy". But,
and I find this confusing, Clone()'s code looks like this:

```
  Local<v8::Object> v8::Object::Clone() {
    auto self = i::Handle<i::JSObject>::cast(Utils::OpenHandle(this));
    auto isolate = self->GetIsolate();
    ENTER_V8_NO_SCRIPT_NO_EXCEPTION(isolate);
    auto result = isolate->factory()->CopyJSObject(self);
    CHECK(!result.is_null());
    return Utils::ToLocal(result);
  }
```

Note the call to CopyJSObject(). From factory.h, that "[r]eturns a
deep copy of the JavaScript object." From factory.cc, the code for
CopyJSObject():

```
  Handle<JSObject> Factory::CopyJSObject(Handle<JSObject> source) {
    return CopyJSObjectWithAllocationSite(source, Handle<AllocationSite>());
  }
```

So, CopyJSObjectWithAllocationSite() is the OG to look at. I won't copy
the code here because it's long, but I will copy one comment that is
important for this project:

```
  // We can only clone regexps, normal objects, api objects, errors or arrays.
  // Copying anything else will break invariants.
```

Again, that's from factory.cc; so, go check it out. That's current as of
this commit; I imagine it's super stable, but if things get funky,
changes to CopyJSObject() and CopyJSOBjectWithAllocationSite() should be
looked at.