tests: added test for multiple args
I want to block folks passing in multiple args when only the first arg
will be deeply cloned. This is to help manage expectations: DeepClone is
meant to be used by iterating over objects rather than either taking in
multiple args or taking in an array of args to clone.